(07 Business Travel) Optimal Calculation
I've added the real world uplift which is to be applied to the registered gCO2/km figures given to each vehicle. This is from the DESNZ methodology paper. It's not clear to me if yearly uplift values apply to the year of the car's registration or to all the fossil fuel vehicles in that year. If it's the latter, then RWuplift would be better placed at the end of the formula, not against each vehicle.

Also a small typo correction from [RC] to [SC] and also [SC] to [OC]